Clean up debugging logging a bit. MSGDEBUG2 now means "print a one-liner for all PML calls into BTL, and also when BTL calls PML with a recv completion (not send completions)" MSGDEBUG1 means print more internal gory detail MSGDEBUG is gone, replaced by MSGDEBUG1
In the process also found that PUT_DEST style fragments could
potentially be leaked in usnic_free() since send_fragment tests were
being applied to see if it was eligible to be freed.

This commit was SVN r29185.

+++ ompi/mca/btl/usnic/btl_usnic_send.h
@@ -231,10 +231,10 @@ ompi_btl_usnic_endpoint_enqueue_frag(
     ompi_btl_usnic_send_frag_t *frag)
 {
 #if MSGDEBUG1
-    opal_output(0, "enq_frag: frag=%p, endpoint=%p, %s, len=%"PRIu64"\n",
+    opal_output(0, "enq_frag: frag=%p, endpoint=%p, %s, len=%lu\n",
             (void*)frag, (void*)endpoint,
             usnic_frag_type(frag->sf_base.uf_type),
-            frag->sf_base.uf_base.des_src->seg_len);
+            (long unsigned)frag->sf_base.uf_base.des_src->seg_len);
     if (frag->sf_base.uf_type == OMPI_BTL_USNIC_FRAG_LARGE_SEND) {
         ompi_btl_usnic_large_send_frag_t *lfrag;
         lfrag = (ompi_btl_usnic_large_send_frag_t *)frag;
